% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenlist.R
\name{tokenlist_jp}
\alias{tokenlist_jp}
\title{Token list for Japanese character}
\usage{
tokenlist_jp(tokens = list())
}
\arguments{
\item{tokens}{list}
}
\description{
Token list for Japanese character
}
\examples{
\dontrun{
tokens <- list(purrr::set_names(c(intToUtf8(c(26519, 27278))),
               c(intToUtf8(c(21517, 35422)))))
tokenlist_jp(tokens)
}
}
